misc server cleanup
