+++ deps/s3.ts
@@ -1,0 +1,1 @@
+export * from "https://deno.land/x/s3@0.5.0/mod.ts"
